refactor(AttendanceManager): System 객체 삭제 및 처리 이동
